replace to separated package Scenario* classes
