+++ src/main/java/executor/service/service/scenarios/ScenarioSourceListener.java
@@ -1,4 +1,4 @@
-package executor.service.service;
+package executor.service.service.scenarios;
 
 import executor.service.handler.ScenarioHandler;
 import executor.service.model.Scenario;
