+++ src/main/java/executor/service/service/scenarios/impl/JSONFileScenarioProvider.java
@@ -1,9 +1,9 @@
-package executor.service.service.impl;
+package executor.service.service.scenarios.impl;
 
 import com.google.common.reflect.TypeToken;
 import com.google.gson.Gson;
 import executor.service.model.Scenario;
-import executor.service.service.ScenarioProvider;
+import executor.service.service.scenarios.ScenarioProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
