+++ src/main/java/executor/service/service/scenarios/impl/ScenarioSourceListenerImpl.java
@@ -1,9 +1,9 @@
-package executor.service.service.impl;
+package executor.service.service.scenarios.impl;
 
 import executor.service.handler.ScenarioHandler;
 import executor.service.model.Scenario;
-import executor.service.service.ScenarioProvider;
-import executor.service.service.ScenarioSourceListener;
+import executor.service.service.scenarios.ScenarioProvider;
+import executor.service.service.scenarios.ScenarioSourceListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,9 +12,10 @@ import java.util.List;
 public class ScenarioSourceListenerImpl implements ScenarioSourceListener {
 
     private static final long DELAY = 1000;
+    private static final Logger log = LoggerFactory.getLogger(ScenarioSourceListener.class);
+
     private final ScenarioProvider provider;
 
-    private static final Logger log = LoggerFactory.getLogger(ScenarioSourceListener.class);
 
     public ScenarioSourceListenerImpl(ScenarioProvider provider) {
         this.provider = provider;
@@ -28,7 +29,7 @@ public class ScenarioSourceListenerImpl implements ScenarioSourceListener {
     }
 
     private void validateScenarios(List<Scenario> scenarios) {
-        if(scenarios == null || scenarios.isEmpty()) {
+        if (scenarios == null || scenarios.isEmpty()) {
             log.error("Bad scenarios list.");
             throw new IllegalArgumentException("List cannot be null or empty");
         }
@@ -36,7 +37,7 @@ public class ScenarioSourceListenerImpl implements ScenarioSourceListener {
 
     private void listen(List<Scenario> scenarioList, ScenarioHandler handler) {
         int currentIndex = 0;
-        while(true){
+        while (true) {
             Scenario scenario = scenarioList.get(currentIndex);
             handler.onScenarioReceived(scenario);
             currentIndex = (currentIndex + 1) % scenarioList.size();
