+++ src/main/java/executor/service/service/stepexecution/impl/ScenarioExecutorImpl.java
@@ -1,7 +1,7 @@
 package executor.service.service.stepexecution.impl;
 
 import executor.service.model.Scenario;
-import executor.service.service.ScenarioExecutor;
+import executor.service.service.scenarios.ScenarioExecutor;
 import org.openqa.selenium.WebDriver;
 
 public class ScenarioExecutorImpl implements ScenarioExecutor {
